feat: cascadingly delete resources depending on group
Tables with foreign keys cannot delete resources with depending resources. Consequently, the options are to either not define foreign keys at all (and delete resources in the processors of the respective resource as a result of events) or define foreign keys with a cascading deletion on DB level. The former was chosen for a broader DB compatibility.