refactor: implement alternative database (#13)
* refactor: replace Postgres by Cockroach

* fix: deployability in actual clusters

The database provided by the Stackgres operator was not stable enough to be used. Therefore, TiDB and Cockroach were evaluated. TiDB offers a limited MySQL syntax and Cockroach a limited Postgres syntax. Both did quite well when deployed with their respective K8s operators. The layered approach by TiDB seemed to be better at scaling individual parts of the application whereas Cockroach is mostly compatible with PostgreSQL. As a future migration to a fully open source alternative based on Postgres is more likely than a migration to a MySQL-based alternative Cockroach is the database which this PR refactors the project to.